Moved deco helpers to api.tools
